--- account_bank_statement_import_txt_xlsx/models/__init__.py
@@ -1,6 +1,0 @@
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
-
-from . import account_bank_statement_import_sheet_mapping
-from . import account_bank_statement_import_sheet_parser
-from . import account_bank_statement_import
-from . import account_journal
