--- account_bank_statement_import_txt_xlsx/readme/DESCRIPTION.rst
@@ -1,2 +1,0 @@
-This module allows you to import the any TXT/CSV or XLSX file in Odoo as bank
-statements.
