--- account_bank_statement_import_txt_xlsx/tests/__init__.py
@@ -1,3 +1,0 @@
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
-
-from . import test_account_bank_statement_import_txt_xlsx
