--- account_bank_statement_import_txt_xlsx/views/account_bank_statement_import.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!--
-    Copyright 2019 ForgeFlow, S.L.
-    Copyright 2020 CorporateHub (https://corporatehub.eu)
-    License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
--->
-<odoo>
-    <record id="account_bank_statement_import_view" model="ir.ui.view">
-        <field name="name">account.bank.statement.import</field>
-        <field name="model">account.bank.statement.import</field>
-        <field
-            name="inherit_id"
-            ref="account_bank_statement_import.account_bank_statement_import_view"
-        />
-        <field name="arch" type="xml">
-            <xpath expr="//ul[@id='statement_format']" position="inside">
-                <li>
-                    TXT/CSV/XLSX mapping: <field name="sheet_mapping_id" nolabel="1" />
-                </li>
-            </xpath>
-        </field>
-    </record>
-</odoo>
