+++ account_statement_import_txt_xlsx/README.rst
@@ -1,0 +1,124 @@
+==================================
+Bank Statement TXT/CSV/XLSX Import
+==================================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fbank--statement--import-lightgray.png?logo=github
+    :target: https://github.com/OCA/bank-statement-import/tree/14.0/account_statement_import_txt_xlsx
+    :alt: OCA/bank-statement-import
+.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
+    :target: https://translation.odoo-community.org/projects/bank-statement-import-14-0/bank-statement-import-14-0-account_statement_import_txt_xlsx
+    :alt: Translate me on Weblate
+.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
+    :target: https://runbot.odoo-community.org/runbot/174/14.0
+    :alt: Try me on Runbot
+
+|badge1| |badge2| |badge3| |badge4| |badge5| 
+
+This module allows you to import any TXT/CSV or XLSX file in Odoo as bank
+statements.
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Configuration
+=============
+
+To create TXT/CSV/XLSX statement sheet columns mapping:
+
+#. Open *Invoicing > Configuration > Accounting > Statement Sheet Mappings*
+#. Create mapping(s) according to your online banking software statement format
+
+Usage
+=====
+
+To use this module, you need to:
+
+#. Get statement in TXT/CSV or XLSX from your online banking software
+#. Go to Odoo and and import the statement file, selecting corresponding format
+
+Changelog
+=========
+
+12.0.2.0.0
+~~~~~~~~~~
+
+* [BREAKING] New mapping, please review mappings after upgrade.
+* [BREAKING] Different bank accounts have to be used per each currency.
+* [ADD] Support for both Statement and Activity reports.
+* [ADD] Separate fee and currency exchange parsing.
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/bank-statement-import/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/OCA/bank-statement-import/issues/new?body=module:%20account_statement_import_txt_xlsx%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* ForgeFlow
+* CorporateHub
+
+Contributors
+~~~~~~~~~~~~
+
+* Alexis de Lattre <alexis.delattre@akretion.com>
+* Sebastien BEAU <sebastien.beau@akretion.com>
+* Tecnativa (https://www.tecnativa.com)
+
+  * Vicent Cubells <vicent.cubells@tecnativa.com>
+  * Victor M.M. Torres <victor.martin@tecnativa.com>
+
+* ForgeFlow (https://www.forgeflow.com)
+
+  * Jordi Ballester Alomar <jordi.ballester@forgeflow.com>
+  * Miquel Raïch Regué <miquel.raich@forgeflow.com>
+
+* `CorporateHub <https://corporatehub.eu/>`__
+
+  * Alexey Pelykh <alexey.pelykh@corphub.eu>
+
+Maintainers
+~~~~~~~~~~~
+
+This module is maintained by the OCA.
+
+.. image:: https://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: https://odoo-community.org
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
+
+.. |maintainer-alexey-pelykh| image:: https://github.com/alexey-pelykh.png?size=40px
+    :target: https://github.com/alexey-pelykh
+    :alt: alexey-pelykh
+
+Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
+
+|maintainer-alexey-pelykh| 
+
+This module is part of the `OCA/bank-statement-import <https://github.com/OCA/bank-statement-import/tree/14.0/account_statement_import_txt_xlsx>`_ project on GitHub.
+
+You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
