+++ account_statement_import_txt_xlsx/data/map_data.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<!--
+    Copyright 2019 ForgeFlow, S.L.
+    Copyright 2020 CorporateHub (https://corporatehub.eu)
+    License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+-->
+<odoo noupdate="1">
+    <record id="sample_statement_map" model="account.statement.import.sheet.mapping">
+        <field name="name">Sample Statement</field>
+        <field name="float_thousands_sep">comma</field>
+        <field name="float_decimal_sep">dot</field>
+        <field name="delimiter">comma</field>
+        <field name="quotechar">"</field>
+        <field name="timestamp_format">%m/%d/%Y</field>
+        <field name="timestamp_column">Date</field>
+        <field name="amount_column">Amount</field>
+        <field name="original_currency_column">Currency</field>
+        <field name="original_amount_column">Amount Currency</field>
+        <field name="description_column">Label</field>
+        <field name="partner_name_column">Partner Name</field>
+        <field name="bank_account_column">Bank Account</field>
+    </record>
+</odoo>
