+++ account_statement_import_txt_xlsx/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This module allows you to import any TXT/CSV or XLSX file in Odoo as bank
+statements.
