+++ account_statement_import_txt_xlsx/tests/fixtures/original_currency.csv
