+++ account_statement_import_txt_xlsx/wizards/__init__.py
@@ -1,0 +1,3 @@
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+from . import account_statement_import_sheet_mapping_wizard
