--- setup/account_bank_statement_import_txt_xlsx/odoo/addons/account_bank_statement_import_txt_xlsx
@@ -1,1 +1,0 @@
-../../../../account_bank_statement_import_txt_xlsx
