+++ setup/account_statement_import_txt_xlsx/odoo/addons/account_statement_import_txt_xlsx
@@ -1,0 +1,1 @@
+../../../../account_statement_import_txt_xlsx
